ai-backend/requirements.txt
Flask
Flask-CORS
transformers
torch
numpy
pandas
scikit-learn
nltk
requests
python-dotenv
gunicorn 